src/evaluate/evaluate_model.py
#!/usr/bin/env python3
"""
Model Evaluation Script for Historia Scribe

This script evaluates a fine-tuned TrOCR model on a test dataset,
computing Character Error Rate (CER) and Word Error Rate (WER).
Based on specifications in Section 5 of the roadmap.
"""

import argparse
import json
import sys
from pathlib import Path
from typing import Dict, List, Any

import torch
from datasets import Dataset
from transformers import TrOCRProcessor, VisionEncoderDecoderModel
from peft import PeftModel
import jiwer

# Add the src directory to the Python path
sys.path.append(str(Path(__file__).parent.parent))


def load_model_and_processor(model_path: Path) -> tuple:
    """
    Load a fine-tuned model and its processor.
    
    Args:
        model_path: Path to the saved model directory
        
    Returns:
        Tuple of (model, processor)
    """
    # Load processor
    processor = TrOCRProcessor.from_pretrained(model_path)
    
    # Load base model
    base_model = VisionEncoderDecoderModel.from_pretrained(
        "microsoft/trocr-large-handwritten"
    )
    
    # Load LoRA adapter
    model = PeftModel.from_pretrained(base_model, model_path)
    
    return model, processor


def load_test_dataset(config: Dict[str, Any]) -> Dataset:
    """
    Load the test dataset for evaluation.
    
    Args:
        config: Configuration dictionary
        
    Returns:
        Test dataset
    """
    # TODO: Implement proper test dataset loading
    # This should load from data/processed/test/
    print("Test dataset loading not yet implemented - using placeholder")
    
    # Placeholder: Create dummy test data
    dummy_data = {
        "pixel_values": [torch.randn(3, 384, 384) for _ in range(5)],
        "text": ["sample text one", "sample text two", "sample text three", 
                 "sample text four", "sample text five"]
    }
    return Dataset.from_dict(dummy_data)


def normalize_text(text: str, config: Dict[str, Any]) -> str:
    """
    Normalize text according to evaluation configuration.
    
    Args:
        text: Input text to normalize
        config: Configuration with normalization settings
        
    Returns:
        Normalized text
    """
    normalization_config = config["evaluation_params"]["text_normalization"]
    
    if normalization_config["lowercase"]:
        text = text.lower()
    
    if normalization_config["collapse_whitespace"]:
        text = " ".join(text.split())
    
    # TODO: Add punctuation removal if configured
    
    return text


def compute_metrics(
    predictions: List[str], 
    references: List[str], 
    config: Dict[str, Any]
) -> Dict[str, float]:
    """
    Compute CER and WER metrics using jiwer library.
    
    Args:
        predictions: List of predicted texts
        references: List of reference (ground truth) texts
        config: Configuration dictionary
        
    Returns:
        Dictionary with CER and WER values
    """
    # Normalize texts
    norm_predictions = [normalize_text(pred, config) for pred in predictions]
    norm_references = [normalize_text(ref, config) for ref in references]
    
    # Compute CER (Character Error Rate)
    cer_transform = jiwer.Compose([
        jiwer.ToLowerCase(),
        jiwer.RemoveWhiteSpace(replace_by_space=True),
        jiwer.RemoveMultipleSpaces(),
        jiwer.Strip(),
    ])
    
    cer = jiwer.cer(
        norm_references,
        norm_predictions,
        truth_transform=cer_transform,
        hypothesis_transform=cer_transform
    )
    
    # Compute WER (Word Error Rate)
    wer_transform = jiwer.Compose([
        jiwer.ToLowerCase(),
        jiwer.RemovePunctuation(),
        jiwer.RemoveMultipleSpaces(),
        jiwer.Strip(),
        jiwer.ReduceToSingleSentence(),
        jiwer.ReduceToListOfListOfWords(),
    ])
    
    wer = jiwer.wer(
        norm_references,
        norm_predictions,
        truth_transform=wer_transform,
        hypothesis_transform=wer_transform
    )
    
    return {
        "cer": cer,
        "wer": wer,
        "samples_evaluated": len(predictions)
    }


def evaluate_model(
    model: VisionEncoderDecoderModel,
    processor: TrOCRProcessor,
    test_dataset: Dataset,
    config: Dict[str, Any]
) -> Dict[str, float]:
    """
    Evaluate the model on the test dataset.
    
    Args:
        model: Fine-tuned TrOCR model
        processor: TrOCR processor
        test_dataset: Test dataset
        config: Configuration dictionary
        
    Returns:
        Dictionary with evaluation metrics
    """
    model.eval()
    
    predictions = []
    references = []
    
    print(f"Evaluating model on {len(test_dataset)} samples...")
    
    for i, sample in enumerate(test_dataset):
        # TODO: Process actual image data
        # For now, we'll use placeholder text
        
        # Placeholder prediction
        predicted_text = "sample predicted text"
        
        # Get reference text (from dataset)
        reference_text = sample.get("text", "sample reference text")
        
        predictions.append(predicted_text)
        references.append(reference_text)
        
        if (i + 1) % 10 == 0:
            print(f"Processed {i + 1}/{len(test_dataset)} samples")
    
    # Compute metrics
    metrics = compute_metrics(predictions, references, config)
    
    return metrics


def main() -> None:
    """Main evaluation function."""
    parser = argparse.ArgumentParser(
        description="Evaluate Historia Scribe model"
    )
    parser.add_argument(
        "model_path",
        type=Path,
        help="Path to the fine-tuned model directory"
    )
    parser.add_argument(
        "--config",
        type=Path,
        default=Path("configs/config.yml"),
        help="Path to configuration file (default: configs/config.yml)"
    )
    parser.add_argument(
        "--output-file",
        type=Path,
        help="Path to save evaluation results (optional)"
    )
    
    args = parser.parse_args()
    
    # Load configuration
    import yaml
    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)
    
    # Load model and processor
    print(f"Loading model from {args.model_path}")
    model, processor = load_model_and_processor(args.model_path)
    
    # Load test dataset
    test_dataset = load_test_dataset(config)
    
    # Evaluate model
    metrics = evaluate_model(model, processor, test_dataset, config)
    
    # Print results
    print("\n" + "="*50)
    print("EVALUATION RESULTS")
    print("="*50)
    print(f"Character Error Rate (CER): {metrics['cer']:.4f}")
    print(f"Word Error Rate (WER): {metrics['wer']:.4f}")
    print(f"Samples evaluated: {metrics['samples_evaluated']}")
    print("="*50)
    
    # Save results if requested
    if args.output_file:
        with open(args.output_file, 'w') as f:
            json.dump(metrics, f, indent=2)
        print(f"Results saved to {args.output_file}")


if __name__ == "__main__":
    main()
